Change the default type of the timeLag function to first.e.decay value
Solve problem with the computation of corrDimensions of q>2. The corrMatrix parameter of the .C function had the wrong name: corr.matrix. As a result, it returned null values.

The getCorrMatrix has been added to the workspace

change the PlotCorrDim function. The axis of the generalized corrDim was log x instead
(q-1)log(x)

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/nonlinear@13 edb9625f-4e0d-4859-8d74-9fd3b1da38cb